Implement an observation matrix class to simplify use of observation matrices
