+++ src/toast/ops/filterbin.py
@@ -21,7 +21,6 @@ from .._libtoast import (
     fourier,
     legendre,
 )
-from ..covariance import covariance_apply, covariance_invert
 from ..mpi import MPI, get_world
 from ..observation import default_values as defaults
 from ..pixels import PixelData, PixelDistribution
@@ -211,225 +210,6 @@ def combine_observation_matrix(rootname):
     log.info_rank(f"All done in", timer=timer0, comm=None)
 
     return f"{rootname}.npz"
-
-
-def coadd_observation_matrix(
-    inmatrix,
-    outmatrix,
-    file_invcov=None,
-    file_cov=None,
-    nside_submap=16,
-    rcond_limit=1e-3,
-    double_precision=False,
-    comm=None,
-):
-    """Co-add noise-weighted observation matrices
-
-    Args:
-        inmatrix(iterable) : One or more noise-weighted observation
-            matrix files.  If a matrix is used to model several similar
-            observations, append `+N` to the file name to indicate the
-             multiplicity.
-        outmatrix(string) : Name of output file
-        file_invcov(string) : Name of output inverse covariance file
-        file_cov(string) : Name of output covariance file
-        nside_submap(int) : Submap size is 12 * nside_submap ** 2.
-            Number of submaps is (nside / nside_submap) ** 2
-        rcond_limit(float) : "Reciprocal condition number limit
-        double_precision(bool) : Output in double precision
-    """
-
-    log = Logger.get()
-    if comm is None:
-        comm, ntask, rank = get_world()
-    else:
-        ntask = comm.size
-        rank = comm.rank
-    timer0 = Timer()
-    timer1 = Timer()
-    timer0.start()
-    timer1.start()
-
-    if double_precision:
-        dtype = np.float64
-    else:
-        dtype = np.float32
-
-    if len(inmatrix) == 1:
-        # Only one file provided, try interpreting it as a text file with a list
-        try:
-            with open(inmatrix[0], "r") as listfile:
-                infiles = listfile.readlines()
-            log.info_rank(f"Loaded {inmatrix[0]} in", timer=timer1, comm=comm)
-        except UnicodeDecodeError:
-            # Didn't work. Assume that user supplied a single matrix file
-            infiles = inmatrix
-    else:
-        infiles = inmatrix
-
-    obs_matrix_sum = None
-    invcov_sum = None
-    nnz = None
-    npix = None
-
-    for ifine, infile_matrix in enumerate(infiles):
-        infile_matrix = infile_matrix.strip()
-        if "noiseweighted" not in infile_matrix:
-            msg = (
-                f"Observation matrix does not seem to be "
-                f"noise-weighted: '{infile_matrix}'"
-            )
-            raise RuntimeError(msg)
-        if "+" in infile_matrix:
-            infile_matrix, N = infile_matrix.split("+")
-            N = float(N)
-        else:
-            N = 1
-        if not os.path.isfile(infile_matrix):
-            msg = f"Matrix not found: {infile_matrix}"
-            raise RuntimeError(msg)
-        prefix = ""
-        log.info(f"{prefix}Loading {infile_matrix}")
-        obs_matrix = scipy.sparse.load_npz(infile_matrix)
-        if obs_matrix.nnz < 0:
-            msg = f"Overflow in {infile_matrix}: nnz = {obs_matrix.nnz}.\n"
-            raise RuntimeError(msg)
-        if N != 1:
-            obs_matrix *= N
-        if obs_matrix_sum is None:
-            obs_matrix_sum = obs_matrix
-        else:
-            obs_matrix_sum += obs_matrix
-        log.info_rank(f"{prefix}Loaded {infile_matrix} in", timer=timer1, comm=None)
-
-        # We'll need the white noise covariance as well
-        infile_invcov = infile_matrix.replace("noiseweighted_obs_matrix.npz", "invcov")
-        if os.path.isfile(infile_invcov + ".fits"):
-            infile_invcov += ".fits"
-        elif os.path.isfile(infile_invcov + ".h5"):
-            infile_invcov += ".h5"
-        else:
-            msg = (
-                f"Cannot find an inverse covariance matrix to go "
-                "with '{infile_matrix}'"
-            )
-            raise RuntimeError(msg)
-        log.info(f"{prefix}Loading {infile_invcov}")
-        invcov = read_healpix(
-            infile_invcov, None, nest=True, dtype=float, verbose=False
-        )
-        if N != 1:
-            invcov *= N
-        if invcov_sum is None:
-            invcov_sum = invcov
-            nnzcov, npix = invcov.shape
-            nnz = 1
-            while (nnz * (nnz + 1)) // 2 != nnzcov:
-                nnz += 1
-            npixtot = npix * nnz
-        else:
-            invcov_sum += invcov
-        log.info_rank(f"{prefix}Loaded {infile_invcov} in", timer=timer1, comm=None)
-
-    # Put the inverse white noise covariance in a TOAST pixel object
-
-    npix_submap = 12 * nside_submap**2
-    nsubmap = npix // npix_submap
-    local_submaps = [submap for submap in range(nsubmap) if submap % ntask == rank]
-    dist = PixelDistribution(
-        n_pix=npix, n_submap=nsubmap, local_submaps=local_submaps, comm=comm
-    )
-    dist_cov = PixelData(dist, float, n_value=nnzcov)
-    for local_submap, global_submap in enumerate(local_submaps):
-        pix_start = global_submap * npix_submap
-        pix_stop = pix_start + npix_submap
-        dist_cov.data[local_submap] = invcov_sum[:, pix_start:pix_stop].T
-    del invcov_sum
-
-    # Optionally write out the inverse white noise covariance
-
-    if file_invcov is not None:
-        log.info_rank(f"Writing {file_invcov}", comm=comm)
-        if filename_is_fits(file_invcov):
-            write_healpix_fits(
-                dist_cov,
-                file_invcov,
-                nest=True,
-                single_precision=not double_precision,
-            )
-        else:
-            write_healpix_hdf5(
-                dist_cov,
-                file_invcov,
-                nest=True,
-                single_precision=not double_precision,
-                force_serial=True,
-            )
-        log.info_rank(f"Wrote {file_invcov}", timer=timer1, comm=comm)
-
-    # Invert the white noise covariance
-
-    log.info_rank("Inverting white noise matrices", comm=comm)
-    dist_rcond = PixelData(dist, float, n_value=1)
-    covariance_invert(dist_cov, rcond_limit, rcond=dist_rcond, use_alltoallv=True)
-    log.info_rank(f"Inverted white noise matrices in", timer=timer1, comm=comm)
-
-    # Optionally write out the white noise covariance
-
-    if file_cov is not None:
-        log.info_rank(f"Writing {file_cov}", comm=comm)
-        if filename_is_fits(file_cov):
-            write_healpix_fits(
-                dist_cov,
-                file_cov,
-                nest=True,
-                single_precision=not double_precision,
-            )
-        else:
-            write_healpix_hdf5(
-                dist_cov,
-                file_cov,
-                nest=True,
-                single_precision=not double_precision,
-                force_serial=True,
-            )
-        log.info_rank(f"Wrote {file_cov} in", timer=timer1, comm=comm)
-
-    # De-weight the observation matrix
-
-    log.info_rank(f"De-weighting obs matrix", comm=comm)
-    cc = scipy.sparse.dok_matrix((npixtot, npixtot), dtype=np.float64)
-    nsubmap = dist_cov.distribution.n_submap
-    npix_submap = dist_cov.distribution.n_pix_submap
-    for isubmap_local, isubmap_global in enumerate(dist_cov.distribution.local_submaps):
-        submap = dist_cov.data[isubmap_local]
-        offset = isubmap_global * npix_submap
-        for pix_local in range(npix_submap):
-            if np.all(submap[pix_local] == 0):
-                continue
-            pix = pix_local + offset
-            icov = 0
-            for inz in range(nnz):
-                for jnz in range(inz, nnz):
-                    cc[pix + inz * npix, pix + jnz * npix] = submap[pix_local, icov]
-                    if inz != jnz:
-                        cc[pix + jnz * npix, pix + inz * npix] = submap[pix_local, icov]
-                    icov += 1
-    cc = cc.tocsr()
-    obs_matrix_sum = cc.dot(obs_matrix_sum)
-    log.info_rank(f"De-weighted obs matrix in", timer=timer1, comm=comm)
-
-    # Write out the co-added and de-weighted matrix
-
-    if not outmatrix.endswith(".npz"):
-        outmatrix += ".npz"
-    log.info_rank(f"Writing {outmatrix}", comm=comm)
-    scipy.sparse.save_npz(outmatrix, obs_matrix_sum.astype(dtype))
-    log.info_rank(f"Wrote {outmatrix} in", timer=timer1, comm=comm)
-
-    log.info_rank(f"Co-added and de-weighted obs matrix in", timer=timer0, comm=comm)
-
-    return outmatrix
 
 
 @trait_docs
